LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/stretch/stretch-block-size-001-expected.txt

CONSOLE MESSAGE: No valid data-* attributes found in selector list : [data-expected-height]








FAIL [data-expected-height] 1 assert_equals:
<div class="test" data-expected-height="45"></div>
height expected 45 but got 10
FAIL [data-expected-height] 2 assert_equals:
<canvas class="test" data-expected-height="45"></canvas>
height expected 45 but got 20
FAIL [data-expected-height] 3 assert_equals:
<input class="test" data-expected-height="45">
height expected 45 but got 23
FAIL [data-expected-height] 4 assert_equals:
<textarea class="test" data-expected-height="45"></textarea>
height expected 45 but got 36
FAIL [data-expected-height] 5 assert_equals:
<button class="test" data-expected-height="45"></button>
height expected 45 but got 10
FAIL [data-expected-height] 6 assert_equals:
<div class="test" style="float: left" data-expected-height="45"></div>
height expected 45 but got 10
PASS [data-expected-height] 7
FAIL [data-expected-height] 8 assert_equals:
<div class="test" style="position: absolute" data-expected-height="50"></div>
height expected 50 but got 10
FAIL [data-expected-height] 9 assert_equals:
<div class="test" style="position: absolute; inset-block-start: 10px" data-expected-height="45"></div>
height expected 45 but got 10
FAIL [data-expected-height] 10 assert_equals:
<div class="test" style="position: absolute; inset-block-end: 10px" data-expected-height="45"></div>
height expected 45 but got 10
PASS [data-expected-height] 11
PASS [data-expected-height] 12
PASS [data-expected-height] 13
FAIL [data-expected-height] 14 assert_equals:
<div class="test" style="position: absolute" data-expected-height="55"></div>
height expected 55 but got 10
FAIL [data-expected-height] 15 assert_equals:
<div class="test" style="position: absolute; inset-block-start: 10px" data-expected-height="45"></div>
height expected 45 but got 10
FAIL [data-expected-height] 16 assert_equals:
<div class="test" style="position: absolute; inset-block-end: 10px" data-expected-height="45"></div>
height expected 45 but got 10
PASS [data-expected-height] 17
PASS [data-expected-height] 18
PASS [data-expected-height] 19
FAIL [data-expected-height] 20 assert_equals:
<div class="test" data-expected-height="45"></div>
height expected 45 but got 10
PASS [data-expected-height] 21
PASS [data-expected-height] 22
PASS [data-expected-height] 23
